Update Icon Background
